+++ abletonosc/__init__.py
@@ -5,13 +5,18 @@ logger.info("Reloading abletonosc...")
 
 from .osc_server import OSCServer
 from .application import ApplicationHandler
+from .application_view import ApplicationViewHandler
 from .song import SongHandler
 from .clip import ClipHandler
 from .clip_slot import ClipSlotHandler
+from .clip_view import ClipViewHandler
 from .track import TrackHandler
+from .track_view import TrackViewHandler
 from .device import DeviceHandler
+from .device_view import DeviceViewHandler
 from .scene import SceneHandler
 from .view import ViewHandler
+from .song_view import SongViewHandler
 from .midimap import MidiMapHandler
 from .session_ring import SessionRingHandler
 from .constants import OSC_LISTEN_PORT, OSC_RESPONSE_PORT